+++ include/geos/geom/LineString.h
@@ -92,7 +92,7 @@ public:
 	/// Returns a read-only pointer to internal CoordinateSequence
 	const CoordinateSequence* getCoordinatesRO() const;
 
-	virtual const Coordinate& getCoordinateN(int n) const;
+	virtual const Coordinate& getCoordinateN(size_t n) const;
 
 	/// Returns line dimension (1)
 	Dimension::DimensionType getDimension() const override;
