+++ include/geos/noding/SingleInteriorIntersectionFinder.h
@@ -99,8 +99,8 @@ public:
 	 * (e.g. by an disjoint envelope test).
 	 */
 	void processIntersections(
-		SegmentString* e0,  int segIndex0,
-		SegmentString* e1,  int segIndex1) override;
+		SegmentString* e0,  size_t segIndex0,
+		SegmentString* e1,  size_t segIndex1) override;
 
 	bool isDone() const override
 	{
