+++ include/geos/operation/buffer/BufferInputLineSimplifier.h
@@ -122,11 +122,11 @@ private:
 	 * @return the next non-deleted index, if any
 	 * @return inputLine.size() if there are no more non-deleted indices
 	 */
-	unsigned int findNextNonDeletedIndex(unsigned int index) const;
+	size_t findNextNonDeletedIndex(size_t index) const;
 
 	std::unique_ptr<geom::CoordinateSequence> collapseLine() const;
 
-	bool isDeletable(int i0, int i1, int i2, double distanceTol) const;
+	bool isDeletable(size_t i0, size_t i1, size_t i2, double distanceTol) const;
 
 	bool isShallowConcavity(const geom::Coordinate& p0,
 	                        const geom::Coordinate& p1,
@@ -148,7 +148,7 @@ private:
 	 */
 	bool isShallowSampled(const geom::Coordinate& p0,
 	                      const geom::Coordinate& p2,
-	                      int i0, int i2, double distanceTol) const;
+	                      size_t i0, size_t i2, double distanceTol) const;
 
 	bool isShallow(const geom::Coordinate& p0,
 	               const geom::Coordinate& p1,
