+++ src/noding/snapround/MCIndexSnapRounder.cpp
@@ -69,7 +69,7 @@ void
 MCIndexSnapRounder::computeVertexSnaps(NodedSegmentString* e)
 {
 	CoordinateSequence& pts0 = *(e->getCoordinates());
-	for (unsigned int i=0, n=static_cast<unsigned int>(pts0.size()-1); i<n; ++i)
+	for (size_t i = 0, n = pts0.size() - 1; i < n; ++i)
 	{
 		HotPixel hotPixel(pts0[i], scaleFactor, li);
 		bool isNodeAdded = pointSnapper->snap(hotPixel, e, i);
